% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtransform.R
\name{backtransform}
\alias{backtransform}
\title{Backtransform scaled predictors}
\usage{
backtransform(data = novelforest$data)
}
\arguments{
\item{data}{Defaults to the data object in \code{novelforestSG}, but could
also be another dataframe with the same predictor names, should you wish to
(back)scale and (back)center using the same means and standard deviations for
any reason.}
}
\value{
A backtransformed dataset with predictors at their original scales.
}
\description{
Backtransform scaled predictors in the \code{novelforestSG} model to their
original scales. This is done by first back-scaling to the log-scale, and
then backtranformed to the original scale (the predictors were
log-transformed as described in Lai et al. 2021).
}
\examples{
# argument could be left blank
backtransform()
}
\references{
Lai, H.R., Tan, G.S.Y., Neo, L., Kee, C.Y., Yee, A.T.K., Tan, H.T.W.
and Chong, K.Y. (2021) Decoupled responses of native and exotic tree
diversities to distance from old-growth forest and soil phosphorous in
novel secondary forests. \emph{Applied Vegetation Science}.
doi: 10.1111/avsc.12548
}
