% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{novelforest}
\alias{novelforest}
\title{Novel forest dataset and model}
\format{
A \code{brms} model output of class \code{brmsfit},
which is a list containing the input \code{data} and other slots
that store the model components. Here, we will only briefly describe
\itemize{
\item how the \code{data} can be extracted and inspected
\item the model summary
}
Please refer to \code{?brms::brmsfit} or \code{?brms::brm} for more information on the
model object.

To access the dataset, use \code{novelforest$data}. The dataset contains the following
response variables:
\describe{
\item{SD_N_0}{first-order native taxonomic diversity, i.e., species richness}
\item{SD_N_2}{second-order native taxonomic diversity, i.e., inverse Simpson index}
\item{SD_E_0}{first-order exotic taxonomic diversity}
\item{SD_E_2}{second-order exotic taxonomic diversity}
\item{FD_N_0}{first-order native functional diversity}
\item{FD_N_2}{second-order native functional diversity}
\item{FD_E_0}{first-order exotic functional diversity}
\item{FD_E_2}{second-order exotic functional diversity,}
}
and the following explanatory variables:
\describe{
\item{dist}{Distance to old-growth forests, in km}
\item{size}{Patch area, in km^2}
\item{nitrogen}{total soil nitrogen, in mg/kg}
\item{phosphorous}{total extractable soil phosphorous, in mg/kg}
\item{potassium}{total extractable soil potassium, in mg/kg}
\item{patch}{Forest patch ID}
}

Note that all explanatory variables were log-transformed and standardised to zero mean and
unit standard deviations. See Lai et al. (2020) for more details on model building and
data collection.
}
\usage{
novelforest
}
\description{
The dataset and model used in Lai et al. (2020), who examined the diversity--environment
relationships of native and exotic species in two types of post-cultivation novel secondary
forests in Singapore. The generalised linear mixed-effect model was fitted via \code{brms::brm} so this package is
required to properly view the model object.
}
\examples{
# To access the dataset
dat <- novelforest$data
summary(dat)

# To summarise the model parameters
summary(novelforest)
}
\references{
Lai HR, Tan CSY, Neo L, Kee CY, Yee ATK, Tan HTW, Chong KY. (2020)
Decoupled responses of native and exotic tree diversities to distance from old-growth forest
and soil phosphorous in novel secondary forests. \emph{Applied Vegetation Science}
}
\seealso{
brms::brmsfit, brms::brm
}
\keyword{datasets}
