% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novelforestSG-package.R
\docType{package}
\name{novelforestSG}
\alias{novelforestSG}
\title{novelforestSG}
\description{
The dataset and model used in Lai et al. (2021)
Decoupled responses of native and exotic tree diversities to
distance from old-growth forest and soil phosphorous in
novel secondary forests. \emph{Applied Vegetation Science}.
}
\details{
The main component of \code{novelforestSG} is the dataset,
which can be accessed with \code{novelforest$data}.
See \code{?novelforest} for more details.
}
\references{
Lai, H.R., Tan, G.S.Y., Neo, L., Kee, C.Y., Yee, A.T.K., Tan, H.T.W.
and Chong, K.Y. (2021) Decoupled responses of native and exotic tree
diversities to distance from old-growth forest and soil phosphorous in
novel secondary forests. \emph{Applied Vegetation Science}, 24, e12548.
https://doi.org/10.1111/avsc.12548
}
\author{
Hao Ran Lai, \email{hrlai.ecology@gmail.com} \cr
Kwek Yan Chong, \email{kwek@u.nus.edu} \cr
Alex Thiam Koon Yee \cr
Maintainer: Hao Ran Lai, \email{hrlai.ecology@gmail.com}
}
